Create README.md
This project includes the software and hardware design for DIY Raspberry Pi based security camera.  The Linux Motion application is used to detect motion and capture images, and additional custom hardware provides infrared illumination at night as well as a light sensor to determine when to enable IR illumination.  The camera used is a 12 MP HQ Raspberry Pi camera with switchable IR filter and a 12 mm Arducam lens.  STL files are included for a 3D printed enclosure.  KiCAD project files are also included for the IR illuminator and light sensor designs.